fix: player icon byte size

+++ src/server/network/protocol/protocolgame.cpp
@@ -4589,7 +4589,7 @@ void ProtocolGame::sendIcons(const std::unordered_set<PlayerIcon> &iconSet, cons
 		msg.add<uint16_t>(static_cast<uint16_t>(icons));
 	} else {
 		// Send as uint32_t in new protocol
-		msg.add<uint32_t>(icons);
+		msg.add<uint64_t>(icons);
 		msg.addByte(enumToValue(iconBakragore)); // Icons Bakragore
 	}
